+++ components/Shared/apiErrorAlert.tsx
@@ -40,7 +40,7 @@ export default function ApiErrorAlert() {
           <Button
             colorScheme="orange"
             color="white"
-            onClick={() => router.replace('/main')}
+            onClick={() => router.replace('/')}
             ml={3}
           >
             메인페이지로 이동
